+++ src/GitHub_Updater/Base.php
@@ -220,7 +220,7 @@ class Base {
 		 * Load hook for shiny updates Basic Authentication headers.
 		 */
 		if ( self::is_doing_ajax() ) {
-			Class_Factory::get_instance( 'Basic_Auth_Loader', self::$options )->load_authentication_hooks();
+			Singleton::get_instance( 'Basic_Auth_Loader', self::$options )->load_authentication_hooks();
 		}
 
 		add_filter( 'extra_theme_headers', array( &$this, 'add_headers' ) );
@@ -244,7 +244,7 @@ class Base {
 		remove_filter( 'http_response', array( 'Fragen\\GitHub_Updater\\API', 'wp_update_response' ) );
 
 		if ( $this->repo_api instanceof Bitbucket_API ) {
-			Class_Factory::get_instance( 'Basic_Auth_Loader', self::$options )->remove_authentication_hooks();
+			Singleton::get_instance( 'Basic_Auth_Loader', self::$options )->remove_authentication_hooks();
 		}
 	}
 
@@ -326,7 +326,7 @@ class Base {
 		if ( self::$load_repo_meta && is_admin() &&
 		     ! apply_filters( 'github_updater_hide_settings', false )
 		) {
-			Class_Factory::get_instance( 'Settings' );
+			Singleton::get_instance( 'Settings' );
 		}
 
 		return true;
@@ -337,7 +337,7 @@ class Base {
 	 */
 	public function ajax_update() {
 		$this->load_options();
-		Class_Factory::get_instance( 'Rest_Update' )->process_request();
+		Singleton::get_instance( 'Rest_Update' )->process_request();
 	}
 
 	/**
@@ -358,7 +358,7 @@ class Base {
 	public function forced_meta_update_plugins( $true = false ) {
 		if ( self::$load_repo_meta || $true ) {
 			$this->load_options();
-			Class_Factory::get_instance( 'Plugin' )->get_remote_plugin_meta();
+			Singleton::get_instance( 'Plugin' )->get_remote_plugin_meta();
 		}
 	}
 
@@ -370,7 +370,7 @@ class Base {
 	public function forced_meta_update_themes( $true = false ) {
 		if ( self::$load_repo_meta || $true ) {
 			$this->load_options();
-			Class_Factory::get_instance( 'Theme' )->get_remote_theme_meta();
+			Singleton::get_instance( 'Theme' )->get_remote_theme_meta();
 		}
 	}
 
@@ -579,7 +579,7 @@ class Base {
 		 * Rename plugins.
 		 */
 		if ( $upgrader instanceof \Plugin_Upgrader ) {
-			$upgrader_object = Class_Factory::get_instance( 'Plugin' );
+			$upgrader_object = Singleton::get_instance( 'Plugin' );
 			if ( isset( $hook_extra['plugin'] ) ) {
 				$slug       = dirname( $hook_extra['plugin'] );
 				$new_source = trailingslashit( $remote_source ) . $slug;
@@ -590,7 +590,7 @@ class Base {
 		 * Rename themes.
 		 */
 		if ( $upgrader instanceof \Theme_Upgrader ) {
-			$upgrader_object = Class_Factory::get_instance( 'Theme' );
+			$upgrader_object = Singleton::get_instance( 'Theme' );
 			if ( isset( $hook_extra['theme'] ) ) {
 				$slug       = $hook_extra['theme'];
 				$new_source = trailingslashit( $remote_source ) . $slug;
@@ -1391,17 +1391,17 @@ class Base {
 			switch ( $transient ) {
 				case 'update_plugins':
 					$this->forced_meta_update_plugins( true );
-					$current = Class_Factory::get_instance( 'Plugin' )->pre_set_site_transient_update_plugins( $current );
+					$current = Singleton::get_instance( 'Plugin' )->pre_set_site_transient_update_plugins( $current );
 					break;
 				case 'update_themes':
 					$this->forced_meta_update_themes( true );
-					$current = Class_Factory::get_instance( 'Theme' )->pre_set_site_transient_update_themes( $current );
+					$current = Singleton::get_instance( 'Theme' )->pre_set_site_transient_update_themes( $current );
 					break;
 				case 'update_core':
 					$this->forced_meta_update_plugins( true );
-					$current = Class_Factory::get_instance( 'Plugin' )->pre_set_site_transient_update_plugins( $current );
+					$current = Singleton::get_instance( 'Plugin' )->pre_set_site_transient_update_plugins( $current );
 					$this->forced_meta_update_themes( true );
-					$current = Class_Factory::get_instance( 'Theme' )->pre_set_site_transient_update_themes( $current );
+					$current = Singleton::get_instance( 'Theme' )->pre_set_site_transient_update_themes( $current );
 					break;
 			}
 			set_site_transient( $transient, $current );