+++ src/GitHub_Updater/Bitbucket_Server_API.php
@@ -490,7 +490,7 @@ class Bitbucket_Server_API extends Bitbucket_API {
 		add_settings_field(
 			'bitbucket_server_username',
 			esc_html__( 'Bitbucket Server Username', 'github-updater' ),
-			array( Class_Factory::get_instance( 'Settings' ), 'token_callback_text' ),
+			array( Singleton::get_instance( 'Settings' ), 'token_callback_text' ),
 			'github_updater_bbserver_install_settings',
 			'bitbucket_server_user',
 			array( 'id' => 'bitbucket_server_username' )
@@ -499,7 +499,7 @@ class Bitbucket_Server_API extends Bitbucket_API {
 		add_settings_field(
 			'bitbucket_server_password',
 			esc_html__( 'Bitbucket Server Password', 'github-updater' ),
-			array( Class_Factory::get_instance( 'Settings' ), 'token_callback_text' ),
+			array( Singleton::get_instance( 'Settings' ), 'token_callback_text' ),
 			'github_updater_bbserver_install_settings',
 			'bitbucket_server_user',
 			array( 'id' => 'bitbucket_server_password', 'token' => true )
@@ -528,7 +528,7 @@ class Bitbucket_Server_API extends Bitbucket_API {
 		$setting_field['page']            = 'github_updater_bbserver_install_settings';
 		$setting_field['section']         = 'bitbucket_server_id';
 		$setting_field['callback_method'] = array(
-			Class_Factory::get_instance( 'Settings' ),
+			Singleton::get_instance( 'Settings' ),
 			'token_callback_checkbox',
 		);
 
@@ -579,7 +579,7 @@ class Bitbucket_Server_API extends Bitbucket_API {
 				parent::$options['bitbucket_server_password'] = $install['bitbucket_password'];
 			}
 
-			Class_Factory::get_instance( 'Bitbucket_Server_API', new \stdClass() );
+			Singleton::get_instance( 'Bitbucket_Server_API', new \stdClass() );
 		}
 
 		return $install;