+++ src/GitHub_Updater/CLI_Integration.php
@@ -34,7 +34,7 @@ class CLI_Integration extends WP_CLI_Command {
 	 * GitHub_Updater_CLI_Integration constructor.
 	 */
 	public function __construct() {
-		$this->base = Class_Factory::get_instance( 'Base' );
+		$this->base = Singleton::get_instance( 'Base' );
 		$this->init_plugins();
 		$this->init_themes();
 	}
@@ -48,7 +48,7 @@ class CLI_Integration extends WP_CLI_Command {
 	public function init_plugins() {
 		$this->base->forced_meta_update_plugins( true );
 		$current = get_site_transient( 'update_plugins' );
-		$current = Class_Factory::get_instance( 'Plugin' )->pre_set_site_transient_update_plugins( $current );
+		$current = Singleton::get_instance( 'Plugin' )->pre_set_site_transient_update_plugins( $current );
 		set_site_transient( 'update_plugins', $current );
 	}
 
@@ -61,7 +61,7 @@ class CLI_Integration extends WP_CLI_Command {
 	public function init_themes() {
 		$this->base->forced_meta_update_themes( true );
 		$current = get_site_transient( 'update_themes' );
-		$current = Class_Factory::get_instance( 'Theme' )->pre_set_site_transient_update_themes( $current );
+		$current = Singleton::get_instance( 'Theme' )->pre_set_site_transient_update_themes( $current );
 		set_site_transient( 'update_themes', $current );
 	}
 
@@ -224,7 +224,7 @@ class CLI_Integration extends WP_CLI_Command {
 		$branch_data['github_updater_branch'] = $cli_config['branch'];
 		$branch_data['repo']                  = $slug;
 
-		Class_Factory::get_instance( 'Branch' )->set_branch_on_install( $branch_data );
+		Singleton::get_instance( 'Branch' )->set_branch_on_install( $branch_data );
 	}
 
 }