+++ src/GitHub_Updater/Theme.php
@@ -347,7 +347,7 @@ class Theme extends Base {
 
 		// Get current branch.
 		$repo   = $this->config[ $theme_key ];
-		$branch = Class_Factory::get_instance( 'Branch' )->get_current_branch( $repo );
+		$branch = Singleton::get_instance( 'Branch' )->get_current_branch( $repo );
 
 		$branch_switch_data                      = array();
 		$branch_switch_data['slug']              = $theme_key;